requirements.txt
python-dotenv==1.1.0
PyQt6==6.9.0
spotipy==2.25.1
SQLAlchemy==2.0.40
wptools==0.4.7
